Shared render so we can reuse rendering components to aid consistent design
