+++ src/queries/me/render.ts
@@ -1,51 +1,13 @@
 import {pipe} from 'fp-ts/lib/function';
-import * as O from 'fp-ts/Option';
-import * as RA from 'fp-ts/ReadonlyArray';
 import {getGravatarThumbnail} from '../../templates/avatar';
-import {
-  html,
-  sanitizeOption,
-  safe,
-  sanitizeString,
-  joinHtml,
-} from '../../types/html';
+import {html, sanitizeOption, safe, sanitizeString} from '../../types/html';
 import {ViewModel} from './view-model';
 import {pageTemplate} from '../../templates';
 import {renderMemberNumber} from '../../templates/member-number';
-import {displayDate} from '../../templates/display-date';
-import {DateTime} from 'luxon';
-
-const howToGetTrained = html`<details>
-  <summary>How to get trained</summary>
-  <div>
-    <ol>
-      <li>
-        Do
-        <a href="https://equipment.makespace.org/">the online training</a>
-      </li>
-      <li>
-        Sign up for practical training session on
-        <a href="https://www.meetup.com/makespace">Meetup</a>
-      </li>
-    </ol>
-    <p>Remember:</p>
-    <ul>
-      <li>All trainers are members volunteering their time.</li>
-      <li>
-        Trainings are often only scheduled on demand after people have passed
-        the online quiz.
-      </li>
-      <li>
-        If you are struggling to join a practical training, reach out to the
-        relevant owners via email.
-      </li>
-      <li>
-        We always need more trainers. If you'd like to help others please let
-        the owners know.
-      </li>
-    </ul>
-  </div>
-</details> `;
+import {renderOwnerAgreementStatus} from '../shared-render/owner-agreement';
+import {renderOwnerStatus} from '../shared-render/owner-status';
+import {renderTrainerStatus} from '../shared-render/trainer-status';
+import {renderTrainingStatus} from '../shared-render/training-status';
 
 const editName = (viewModel: ViewModel) =>
   html`<a href="/members/edit-name?member=${viewModel.member.memberNumber}"
@@ -95,112 +57,6 @@ const renderMemberDetails = (viewModel: ViewModel) => html`
     </tbody>
   </table>
 `;
-
-export const renderOwnerStatus = (ownerOf: ViewModel['member']['ownerOf']) =>
-  pipe(
-    ownerOf,
-    RA.map(
-      area =>
-        html`<li>
-          <a href="/equipment#${safe(area.id)}">${sanitizeString(area.name)}</a>
-          (since ${displayDate(DateTime.fromJSDate(area.ownershipRecordedAt))})
-        </li>`
-    ),
-    RA.match(
-      () => html`
-        <p>You currently do not own any areas.</p>
-        <p>
-          Owners are the members who maintain and expand the capabilities of our
-          areas. We are always looking for more.
-        </p>
-      `,
-      listItems => html`
-        <p>You are an owner of the following areas:</p>
-        <ul>
-          ${joinHtml(listItems)}
-        </ul>
-      `
-    )
-  );
-
-export const renderTrainerStatus = (
-  trainerFor: ViewModel['member']['trainerFor']
-) =>
-  pipe(
-    trainerFor,
-    RA.map(
-      equipment =>
-        html`<li>
-          <a href="/equipment/${equipment.equipment_id}"
-            >${sanitizeString(equipment.equipment_name)}</a
-          >
-          (since ${displayDate(DateTime.fromJSDate(equipment.since))})
-        </li>`
-    ),
-    RA.match(
-      () => html``,
-      listItems => html`
-        <p>You are a trainer for:</p>
-        <ul>
-          ${joinHtml(listItems)}
-        </ul>
-      `
-    )
-  );
-
-export const renderTrainingStatus = (
-  trainedOn: ViewModel['member']['trainedOn']
-) =>
-  pipe(
-    trainedOn,
-    RA.map(
-      equipment =>
-        html`<li>
-          <a href="/equipment/${safe(equipment.id)}"
-            >${sanitizeString(equipment.name)}</a
-          >
-          (since ${displayDate(DateTime.fromJSDate(equipment.trainedAt))})
-        </li>`
-    ),
-    RA.match(
-      () => html`
-        <p>You are currently not allowed to use any RED equipment.</p>
-        ${howToGetTrained}
-      `,
-      listItems => html`
-        <p>You are permitted to use the following RED equipment:</p>
-        <ul>
-          ${joinHtml(listItems)}
-        </ul>
-        ${howToGetTrained}
-      `
-    )
-  );
-
-export const renderOwnerAgreementStatus = (
-  status: ViewModel['member']['agreementSigned'],
-  third_person: boolean
-) =>
-  pipe(
-    status,
-    O.matchW(
-      () => safe(''),
-      date =>
-        third_person
-          ? html`
-              <p>
-                User signed Owners Agreement
-                (${displayDate(DateTime.fromJSDate(date))})
-              </p>
-            `
-          : html`
-              <p>
-                You have signed the Owners Agreement
-                (${displayDate(DateTime.fromJSDate(date))})
-              </p>
-            `
-    )
-  );
 
 const superUserNav = html`
   <h2>Admin</h2>
